Reduce `write_interface` macro code
