+++ flutter_ffi_plugin/example/native/hub/Cargo.toml
@@ -14,6 +14,7 @@ crate-type = ["lib", "cdylib", "staticlib"]
 [dependencies]
 rinf = "6.11.1"
 prost = "0.12.3"
+# tokio = { version = "1", features = ["sync"] }
 wasm-bindgen = "0.2.92"                     # Uncomment this line to target the web
 tokio_with_wasm = "0.4.4"                   # Uncomment this line to target the web
 sample_crate = { path = "../sample_crate" }
