+++ rust_crate/Cargo.toml
@@ -15,7 +15,9 @@ protoc-prebuilt = "0.3.0"
 home = "0.5.9"
 which = "6.0.0"
 allo-isolate = "0.1.24"
+tokio = { version = "1", features = ["rt-multi-thread"] }
 
 [target.'cfg(target_family = "wasm")'.dependencies]
 js-sys = "0.3.69"
 wasm-bindgen = "0.2.92"
+wasm-bindgen-futures = "0.4.42"
