+++ rust_crate/src/interface.rs
@@ -1,3 +1,5 @@
+use std::future::Future;
+
 #[cfg(not(target_family = "wasm"))]
 use super::interface_os::*;
 #[cfg(target_family = "wasm")]
@@ -15,4 +17,11 @@ pub struct DartSignal<T> {
 /// Send a signal to Dart.
 pub fn send_rust_signal(message_id: i32, message_bytes: Vec<u8>, binary: Vec<u8>) {
     send_rust_signal_extern(message_id, message_bytes, binary);
+}
+
+pub fn start_rust_logic<F>(main_future: F)
+where
+    F: Future<Output = ()> + Send + 'static,
+{
+    start_rust_logic_extern(main_future);
 }
